+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.2.0
+
+### Breaking
+
+- Renamed `gln_tVoronoiOpts -> gln_tVoronoiOpts`
+- Moved `gln_tVoronoiOpts` from the `Common` to the `Worley` chunk. 
+
+### Changed
+
+- Added Gerstner Waves
+
 ## Version 1.1.0
 
 ### Breaking